Eventual connectivity and cache for add post
Contingent connectivity and a caching strategy are implemented to allow users to post offline. This is done so that when the user tries to post offline, a pop-up notifies them that they are offline, and when they reconnect, the posts they have in their queue will be synced. Preliminary tests show it works.